Move triggers.ts to common folder
