+++ packages/headless/src/api/common/trigger.ts
