+++ packages/headless/src/ssr.index.ts
@@ -411,7 +411,7 @@ export type {
 } from './controllers/triggers/headless-execute-trigger.ssr';
 export {defineExecuteTrigger} from './controllers/triggers/headless-execute-trigger.ssr';
 
-export type {ExecuteTriggerParams} from './api/search/trigger';
+export type {ExecuteTriggerParams} from './api/common/trigger';
 
 export type {
   NotifyTrigger,
